Refactor Colloscope Models
