+++ packages/polytechcolloscopeclient/lib/polytechcolloscopeclient.dart
@@ -1,3 +1,5 @@
-export 'src/colloscope_datastructs.dart';
 export 'src/consts.dart';
+export 'src/models/StudentColloscope.dart';
+export 'src/models/kholle.dart';
+export 'src/models/student.dart';
 export 'src/polytechcolloscopeclient_base.dart';
